% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biometric_mask.R
\name{biometric_mask}
\alias{biometric_mask}
\alias{ct_biometric_mask}
\alias{mri_biometric_mask}
\title{Estimate Biometric Masks}
\usage{
biometric_mask(file, type = c("ct", "mri"), ...)

ct_biometric_mask(file, ...)

mri_biometric_mask(file, ...)
}
\arguments{
\item{file}{Filename (or nifti) of image}

\item{type}{character of \code{ct} for CT images or \code{mri} for magnetic
resonance images}

\item{...}{arguments to pass to \code{ct/mri_ear_mask} and
\code{ct/mri_face_mask}}
}
\value{
Object of class nifti
}
\description{
Estimates biometric masks for anonymization, such as the face and
ears.
}
\examples{
\dontrun{
file = "~/Desktop/Desktop/scratch/100-318_20070723_0957_CT_3_CT_Head-.nii.gz"
mask = NULL
robust = FALSE
mask = ct_biometric_mask(
   file = file,
   robust = FALSE
   )
 img = readnii(file)
 rimg = randomize_mask(img, mask = face)
}

}
