% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ct_head_mask.R
\name{create_ct_head_mask}
\alias{create_ct_head_mask}
\alias{segment_ct_head}
\title{Create CT Head Mask}
\usage{
create_ct_head_mask(img, fill_size = 11)

segment_ct_head(img, fill_size = 11)
}
\arguments{
\item{img}{CT image of class \code{\link{nifti}}}

\item{fill_size}{Size of fill (in voxels),
passed to \code{\link[extrantsr]{filler}}}
}
\value{
An object of class \code{\link{nifti}}
}
\description{
Create CT Head Mask
}
