% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_dcm2nii.R
\name{ct_dcm2nii}
\alias{ct_dcm2nii}
\title{CT DICOM to NIfTI conversion}
\usage{
ct_dcm2nii(
  basedir = ".",
  merge_files = TRUE,
  verbose = TRUE,
  drop_dim = TRUE,
  ...,
  ignore_roi_if_multiple = FALSE,
  fail_on_error = FALSE,
  uncorrected = FALSE
)
}
\arguments{
\item{basedir}{(character) directory to get files}

\item{merge_files}{Should files be merged, passed do \code{\link{dcm2nii}}
options}

\item{verbose}{print diagnostic messages}

\item{drop_dim}{passed to \code{\link{readnii}} for dropping empty
dimensions}

\item{...}{Additional parameters passed to \code{\link{dcm2nii}}}

\item{ignore_roi_if_multiple}{additional argument to pass to
[dcm2niir::check_dcm2nii()] to remove ROI overlays if present}

\item{fail_on_error}{if the exit code from [dcm2niir::dcm2nii] is not
`0` and `fail_on_error = TRUE`, function will error out}

\item{uncorrected}{passed to [dcm2niir::check_dcm2nii()] to grab the
"uncorrected" scan.  Do not use unless you understand the `dcm2niix`
correction process.}
}
\value{
A list or singular \code{nifti} image
}
\description{
CT DICOM to NIfTI conversion
}
