% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_Skull_Strip.R
\name{CT_Skull_Strip}
\alias{CT_Skull_Strip}
\title{CT Skull Stripping within R}
\usage{
CT_Skull_Strip(img, outfile = NULL, keepmask = TRUE, maskfile = NULL,
  inskull_mesh = FALSE, retimg = TRUE, reorient = FALSE, intern = TRUE,
  betcmd = "bet2", opts = "-f 0.01 -v", presmooth = TRUE, remask = TRUE,
  refill = FALSE, refill.thresh = 0.75, sigma = 1, lthresh = 0,
  uthresh = 100, verbose = TRUE, ...)
}
\arguments{
\item{img}{(character) File to be skull stripped or object of class
nifti}

\item{outfile}{(character) output filename}

\item{keepmask}{(logical) Should we keep the mask?}

\item{maskfile}{(character) Filename for mask (if \code{keepmask = TRUE}).
If \code{NULL}, then will do \code{paste0(outfile, "_Mask")}.}

\item{inskull_mesh}{(logical) Create inskull_mesh file from bet?
(Warning - will take longer)
This an exterior surface of the brain. (experimental)
Also, if \code{outfile} is \code{NULL}, then this will be created in
a temporary directory and not be retrieved.}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{intern}{(logical) pass to \code{\link{system}}}

\item{betcmd}{(character) bet command to be used, see \code{\link{fslbet}}}

\item{opts}{(character) additional options to \code{\link{fslbet}}}

\item{presmooth}{(logical) indicator if pre-smoothing should be
done before BET}

\item{remask}{(logical) Mask the smoothed image with HU mask from initial
step?}

\item{refill}{(logical) indicator to post-smooth mask and then fill}

\item{refill.thresh}{(numeric) Value to threshold post-smoothed mask}

\item{sigma}{(integer) size of Gaussian kernel passed to
\code{\link{fslsmooth}} if \code{presmooth} is \code{TRUE}}

\item{lthresh}{(default: 0) Lower value to threshold CT
\code{\link{fslthresh}}}

\item{uthresh}{(default: 100) Upper value to threshold CT
\code{\link{fslthresh}}}

\item{verbose}{(logical) Should diagnostic output be printed?}

\item{...}{additional arguments passed to \code{\link{fslbet}}.}
}
\value{
character or logical depending on intern
}
\description{
Skull Stripping (using FSL's BET) a CT file using \code{fslr}
functions
}

