% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_template.R
\name{ct_template}
\alias{ct_template}
\title{Read in CT Template}
\usage{
ct_template(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{ct_template_fname}}}
}
\value{
Object of class \code{\link{nifti}}
}
\description{
Wrapper for returning a \code{nifti} object of the CT Template
}

