% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ear_mask.R
\name{ear_mask}
\alias{ct_ear_mask}
\alias{ct_ear_mask,mri_ear_mask}
\alias{ear_mask}
\alias{mri_ear_mask}
\title{Create Mask of the Ears}
\usage{
ear_mask(file, type = c("ct", "mri"),
  template.file = system.file("scct_unsmooth_SS_0.01.nii.gz", package =
  "ichseg"), template.ear_mask = NULL,
  template.left_ear_inds = list(170:180, 60:110, 1:60),
  template.right_ear_inds = list(1:10, 60:110, 1:60), extend_mask = FALSE,
  ...)

ct_ear_mask(...)

mri_ear_mask(..., template.file = mni_fname(brain = TRUE))
}
\arguments{
\item{file}{File for ear masking - either filename or class nifti}

\item{type}{is it a CT or MRI?  Will determine if \code{\link{ct_face_mask}} or
\code{\link{mri_face_mask}} is called}

\item{template.file}{Template to warp to original image space}

\item{template.ear_mask}{Mask of template to use as rough ear mask.  If
\code{template.file} is not specified, \code{template.left_ear_inds} and
\code{template.right_ear_inds}
must be}

\item{template.left_ear_inds}{List of length 3 for indices of
\code{template.file} to indicate the left-ear mask.}

\item{template.right_ear_inds}{List of length 3 for indices of
\code{template.file} to indicate the right-ear mask.}

\item{extend_mask}{passed to \code{\link{ct_face_mask}} or
\code{\link{mri_face_mask}}, but should be \code{FALSE} otherwise the
mask will be extended forward}

\item{...}{arguments passed to \code{\link{ct_face_mask}} or
\code{\link{mri_face_mask}}}
}
\value{
Object of class nifti
}
\description{
Creates a rough mask of the ear from a head scan
}
\examples{
\dontrun{
file = "~/Desktop/Desktop/scratch/100-318_20070723_0957_CT_3_CT_Head-.nii.gz"
mask = NULL
robust = FALSE
ears = ct_ear_mask(
   file = file,
   robust = FALSE
 )
 img = readnii(file)
 rimg = randomize_mask(img, mask = ears)
}
}

