% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_mask.R
\name{face_mask}
\alias{face_mask}
\title{Create Mask of the Face}
\usage{
face_mask(file, mask = NULL, robust = TRUE,
  template.file = system.file("scct_unsmooth_SS_0.01.nii.gz", package =
  "ichseg"), template.mask = NULL, template.mask_inds = list(50:130,
  170:217, 1:15), typeofTransform = "Affine", swapdim = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{file}{File for face masking - either filename or class nifti}

\item{mask}{file or \code{nifti} to mask the \code{file}}

\item{robust}{If \code{mask = NULL}, then \code{robust} is
passed to \code{\link{CT_Skull_Stripper}}}

\item{template.file}{Template to warp to original image space}

\item{template.mask}{Mask of template to use as rough face mask.  If
\code{template.file} is not specified, \code{template.mask_inds}
must be}

\item{template.mask_inds}{List of length 3 for indices of
\code{template.file} to indicate the mask.}

\item{typeofTransform}{Transformation for template to image, passed to
\code{\link{ants_regwrite}}.}

\item{swapdim}{Should the dimensions be swapped before registration,
and then reset after}

\item{verbose}{Print out diagnostic messages}

\item{...}{not used}
}
\value{
Object of class nifti
}
\description{
Creates a rough mask of the face of a brain image
}

