% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ich_candidate_voxels.R
\name{ich_candidate_voxels}
\alias{ich_candidate_voxels}
\title{Create ICH Segmentation Candidate Voxesl}
\usage{
ich_candidate_voxels(df)
}
\arguments{
\item{df}{\code{data.frame} of image from \code{\link{make_predictors}}}
}
\value{
Logical vector
}
\description{
Takes estimated cutoffs from paper from training data
and creates a logical candidate vector
}
