% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ich_predict.R
\name{ich_predict}
\alias{ich_predict}
\title{Predict ICH Images}
\usage{
ich_predict(df, nim, model = c("rf", "logistic", "big_rf"),
  verbose = TRUE, native = TRUE, native_img = NULL,
  transformlist = NULL, interpolator = NULL, native_thresh = 0.5,
  shiny = FALSE, model_list = NULL, smoothed_cutoffs = NULL,
  outfile = NULL, ...)
}
\arguments{
\item{df}{\code{\link{data.frame}} of predictors.  If \code{multiplier}
column does not exist, then \code{\link{ich_candidate_voxels}} will
be called}

\item{nim}{object of class \code{\link{nifti}}, from
\code{\link{make_predictors}}}

\item{model}{model to use for prediction,
either the random forest (rf) or logistic}

\item{verbose}{Print diagnostic output}

\item{native}{Should native-space predictions be given?}

\item{native_img}{object of class \code{\link{nifti}}, which
is the dimensions of the native image}

\item{transformlist}{Transforms list for the transformations back to native space.
NOTE: these will be inverted.}

\item{interpolator}{Interpolator for the transformation back to native space}

\item{native_thresh}{Threshold for re-thresholding binary mask after
interpolation}

\item{shiny}{Should shiny progress be called?}

\item{model_list}{list of model objects, used mainly for retraining
but only expert use.}

\item{smoothed_cutoffs}{A list with an element
\code{mod.dice.coef}, only expert use.}

\item{outfile}{filename for output file.
We write the smoothed, thresholded image.  If \code{native = TRUE},
then the file will be native space, otherwise in registered
space}

\item{...}{Additional options passsed to \code{\link{ich_preprocess}}}
}
\value{
List of output registered and native space
prediction/probability images
}
\description{
This function will take the data.frame of predictors and
predict the ICH voxels from the model chosen.
}
\seealso{
\code{\link{ich_candidate_voxels}}
}
