% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ich_preprocess.R
\name{ich_preprocess}
\alias{ich_preprocess}
\title{Preprocess data for ICH Segmentation}
\usage{
ich_preprocess(img, skull_strip = TRUE, robust = TRUE, mask = NULL,
  template.file = system.file("scct_unsmooth_SS_0.01.nii.gz", package =
  "ichseg"), outprefix = NULL, typeofTransform = c("Rigid", "Affine"),
  interpolator = "Linear", outfile = NULL, verbose = TRUE, ...)
}
\arguments{
\item{img}{CT image, object of class \code{nifti} or
character filename}

\item{skull_strip}{Should the image be skull stripped? If not,
a mask must be specified.  If mask specified, will override this
option}

\item{robust}{If skull stripping, should
\code{\link{CT_Skull_Strip_robust}}
be used vs. \code{\link{CT_Skull_Strip}}}

\item{mask}{binary brain mask, object of class \code{nifti} or
character filename}

\item{template.file}{Template to register to (Skull stripped template)}

\item{outprefix}{Passed to \code{\link[extrantsr]{registration}} for
the output transformations}

\item{typeofTransform}{Transformation to the template}

\item{interpolator}{Interpolation to be done after transformation}

\item{outfile}{Output filename for transformed registered image}

\item{verbose}{Print diagnostic output}

\item{...}{Additional options passsed to \code{\link{CT_Skull_Strip_robust}}
or \code{\link{CT_Skull_Strip}}}
}
\value{
List of output images and transformations
}
\description{
Will do skull stripping and registration to the CT template
}
