% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ich_segment.R
\name{ich_segment}
\alias{ich_segment}
\alias{ich_process_predictors}
\title{Predict ICH Segmentation}
\usage{
ich_segment(
  img,
  ...,
  verbose = TRUE,
  shiny = FALSE,
  model = c("rf", "logistic", "big_rf"),
  outfile = NULL
)

ich_process_predictors(
  img,
  mask = NULL,
  save_imgs = FALSE,
  outdir = NULL,
  stub = NULL,
  verbose = TRUE,
  shiny = FALSE,
  roi = NULL,
  erode_mask = TRUE,
  ...
)
}
\arguments{
\item{img}{CT image, object of class \code{nifti} or
character filename}

\item{...}{Additional options passsed to \code{\link{ich_preprocess}}}

\item{verbose}{Print diagnostic output}

\item{shiny}{Should shiny progress be called?}

\item{model}{model to use for prediction,
either the random forest (rf) or logistic}

\item{outfile}{filename for native-space, smoothed,
thresholded image.}

\item{mask}{binary brain mask, object of class \code{nifti} or
character filename}

\item{save_imgs}{Logical to save all images that are created as
predictors}

\item{outdir}{Output directory of saved images, needs to be set
if \code{save_imgs = TRUE}}

\item{stub}{Basename to write image names if \code{save_imgs = TRUE}}

\item{roi}{Filename of ROI, which will be transformed}

\item{erode_mask}{Should the brain mask be eroded?}
}
\value{
List of output prediction/probability images
}
\description{
Will preprocess and predict the ICH voxels
}
