% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ich_train_data.R
\name{ich_train_data}
\alias{ich_train_data}
\title{Create ICH Train Data}
\usage{
ich_train_data(
  imgs,
  run_candidate = TRUE,
  cutoffs = ichseg::est.cutoffs,
  rois = NULL,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{imgs}{CT image, object of class \code{nifti} (list) or
character filename}

\item{run_candidate}{Run \code{\link{ich_candidate_voxels}}
to get candidate voxels}

\item{cutoffs}{cutoffs to pass to \code{\link{ich_candidate_voxels}}}

\item{rois}{ROI image, object of class \code{nifti} (list) or
character filename}

\item{...}{Additional options passsed to \code{\link{ich_preprocess}}}

\item{verbose}{Print diagnostic output}
}
\value{
List of one \code{data.frame} with the predictors
}
\description{
Create ICH Train Data
}
\note{
This is a simple wrapper for \code{\link{ich_process_predictors}}
}
