% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ich_train_data.R
\name{ich_train_model}
\alias{ich_train_model}
\title{ICH Train Model}
\usage{
ich_train_model(df, do.trace = TRUE, ...)
}
\arguments{
\item{df}{A \code{data.frame} from \code{\link{ich_train_data}} or
the \code{$img.pred$df} output from  \code{\link{ich_process_predictors}}
output list.}

\item{do.trace}{passed to \code{\link{randomForest}} for training verbosity}

\item{...}{additional arguments passed to \code{\link{randomForest}}}
}
\value{
A \code{randomForest} object
}
\description{
ICH Train Model
}
