% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_modlist.R
\docType{data}
\name{logistic_modlist}
\alias{logistic_modlist}
\title{Estimated Logistic Model List for ICH}
\format{A list with 7 elements, which are:
\describe{
\item{mod.pauc}{the partial AUC estimated with the training data }
\item{mod.pauc.cutoff}{the sens/spec for partial AUC cutoff }
\item{mod.sens.cutoff}{the sensitivity and cutoff estimated with the training datan }
\item{mod.dice.cutoff}{the Dice Similarity Index and cutoff estimated with the training data }
\item{mod.acc}{the accuracy and cutoff estimated with the training data}
\item{mod}{the model object of class \code{glm} }
\item{mod.time}{the time it took to train the model }
}}
\usage{
logistic_modlist
}
\description{
A list containing the logistic model for predicting ICH and other factors for
prediction
}
\keyword{datasets}

