% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amoment.R
\name{mean_image}
\alias{mean_image}
\title{Calculate Mean Image by FFT}
\usage{
mean_image(x, nvoxels, shift = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{3D array}

\item{nvoxels}{window (in voxels) for the neighborhood 
(1 results in a 3x3 cube)}

\item{shift}{Should results be shifted back?}

\item{verbose}{Diagnostic outputing}
}
\value{
Array of size of x
}
\description{
This function calculates the mean image using fft
}

