% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_modlist.R
\docType{data}
\name{rf_modlist}
\alias{rf_modlist}
\alias{big_rf_modlist}
\title{Estimated Random Forest Model List for ICH}
\format{A list with 7 elements, which are:
\describe{
\item{mod.pauc}{the partial AUC estimated with the training data }
\item{mod.pauc.cutoff}{the sens/spec for partial AUC cutoff }
\item{mod.sens.cutoff}{the sensitivity and cutoff estimated with the training datan }
\item{mod.dice.cutoff}{the Dice Similarity Index and cutoff estimated with the training data }
\item{mod.acc}{the accuracy and cutoff estimated with the training data}
\item{mod}{the model object of class \code{randomForest} }
\item{mod.time}{the time it took to train the model }
}}
\usage{
rf_modlist

big_rf_modlist
}
\description{
A list containing the random forest model for predicting ICH and other factors for
prediction
}
\keyword{datasets}
