% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_dcm2nii.R
\name{ct_dcm2nii}
\alias{ct_dcm2nii}
\title{CT DICOM to NIfTI conversion}
\usage{
ct_dcm2nii(basedir = ".", merge_files = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{basedir}{(character) directory to get files}

\item{merge_files}{Should files be merged, passed do \code{\link{dcm2nii}}
options}

\item{verbose}{print diagnostic messages}

\item{...}{Additional parameters passed to \code{\link{dcm2nii}}}
}
\value{
A list or singular \code{nifti} image
}
\description{
CT DICOM to NIfTI conversion
}
