% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_template.R
\name{ct_template_fname}
\alias{ct_template_fname}
\title{Get CT Template Filename}
\usage{
ct_template_fname(type = c("brain", "mask", "image"))
}
\arguments{
\item{type}{Type of image.  \code{image} is the original image,
\code{brain} is the skull stripped image,
\code{mask} is brain mask.}
}
\value{
Character filename
}
\description{
Wrapper for returning the filename of the CT Template
}

