% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ear_mask.R
\name{ct_ear_mask}
\alias{ct_ear_mask}
\alias{ct_ear_mask,mri_ear_mask}
\title{Create Mask of the Ears}
\usage{
ct_ear_mask(file, template.file = system.file("scct_unsmooth_SS_0.01.nii.gz",
  package = "ichseg"), template.ear_mask = NULL,
  template.left_ear_inds = list(170:180, 60:110, 1:60),
  template.right_ear_inds = list(1:10, 60:110, 1:60), ...)
}
\arguments{
\item{file}{File for ear masking - either filename or class nifti}

\item{template.file}{Template to warp to original image space}

\item{template.ear_mask}{Mask of template to use as rough ear mask.  If
\code{template.file} is not specified, \code{template.left_ear_inds} and
\code{template.right_ear_inds}
must be}

\item{template.left_ear_inds}{List of length 3 for indices of
\code{template.file} to indicate the left-ear mask.}

\item{template.right_ear_inds}{List of length 3 for indices of
\code{template.file} to indicate the right-ear mask.}

\item{...}{arguments passed to \code{\link{ct_face_mask}}}

\item{mask}{file or \code{nifti} to mask the \code{file}}

\item{robust}{If \code{mask = NULL}, then \code{robust} is
passed to \code{\link{CT_Skull_Stripper}}}
}
\value{
Object of class nifti
}
\description{
Creates a rough mask of the ear from a head scan
}
\examples{
\dontrun{
file = "~/Desktop/Desktop/scratch/100-318_20070723_0957_CT_3_CT_Head-.nii.gz"
mask = NULL
robust = FALSE
face = ct_ear_mask(
   file = file,
   robust = FALSE,
    template.mask = system.file("scct_unsmooth_SS_0.01_Mask.nii.gz",
                  package = "ichseg")
   )
 img = readnii(file)
 rimg = randomize_mask(img, mask = face)
}
#'
}

