% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ich_segment.R
\name{ich_segment}
\alias{ich_segment}
\title{Predict ICH Segmentation}
\usage{
ich_segment(img, mask = NULL, model = c("rf", "logistic"),
  save_imgs = FALSE, outdir = NULL, stub = NULL, verbose = TRUE, ...)
}
\arguments{
\item{img}{CT image, object of class \code{nifti} or
character filename}

\item{mask}{binary brain mask, object of class \code{nifti} or
character filename}

\item{model}{model to use for prediction,
either the random forest (rf) or logistic}

\item{save_imgs}{Logical to save all images that are created as
predictors}

\item{outdir}{Output directory of saved images, needs to be set
if \code{save_imgs = TRUE}}

\item{stub}{Basename to write image names if \code{save_imgs = TRUE}}

\item{verbose}{Print diagnostic output}

\item{...}{Additional options passsed to \code{\link{ich_preprocess}}}
}
\value{
List of output prediction/probability images
}
\description{
Will preprocess and predict the ICH voxels
}

